% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cat.R
\name{gen_cat}
\alias{gen_cat}
\title{Generate a Categorical Variable Object}
\usage{
gen_cat(var_name, var_levels)
}
\arguments{
\item{var_name}{A character string specifying the name of the variable.}

\item{var_levels}{A vector of levels for the categorical variable.}
}
\value{
A categorical variable with the specified levels.
}
\description{
This function generates a categorical variable object with specified levels.
}
\examples{
gen_cat("gender", c("male", "female"))
bmi <- gen_cat("BMI", c("Underweight", "Normal", "Overweight", "Obese"))
attributes(bmi)
}
