% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_grp.R
\name{gen_grp}
\alias{gen_grp}
\title{Generate Group Variable}
\usage{
gen_grp(var_name, var_levels, units = NULL)
}
\arguments{
\item{var_name}{A character vector specifying the name(s) of the variable(s).}

\item{var_levels}{A character vector specifying the levels of the variable.}

\item{units}{Optional. A character vector specifying the units of the variable.}
}
\value{
A group variable with the specified properties.
}
\description{
This function creates a group variable with specified levels and optional units.
}
\examples{
gen_grp(var_name = "Group", var_levels = c("A", "B", "C"))
gen_grp(var_name = "Temperature", var_levels = c("Low", "Medium", "High"), units = "Celsius")
arms <- gen_grp("Treatment", c("SAR1234","ABC", "Placebo"), rep("mg", 2))
attributes(arms)
class(arms)
}
