% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mtx.R
\name{gen_mtx}
\alias{gen_mtx}
\title{Generate a Table Matrix}
\usage{
gen_mtx(..., col_groups = NULL)
}
\arguments{
\item{...}{sections to be put in the table}

\item{col_groups}{A vector or list specifying header group. Defalt is \code{NULL}, which will leave table header empty.}
}
\value{
A matrix object generated based on the input parameters.
}
\description{
This function generates a matrix based on the provided aruments.
}
