% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_list.R
\name{gen_plain_list}
\alias{gen_plain_list}
\title{Generate a Plain List Object}
\usage{
gen_plain_list(
  var_name = "Patient",
  n,
  num_cols = NULL,
  cat_cols = NULL,
  date_cols = NULL
)
}
\description{
Generate a Plain List Object
}
