% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr.R
\name{get_attr}
\alias{get_attr}
\title{Get Attribute from an Object}
\usage{
get_attr(x)
}
\arguments{
\item{x}{An object from which the attribute will be retrieved.}
}
\value{
The value of the attribute if it exists, or \code{NULL} if the attribute is not found.
}
\description{
This is cloture function to generate a function that retrieves a specific attribute from a given object.
}
\examples{
obj <- structure(1:5, my_attr = "example")
get_attr(obj) # Retrieves the "my_attr" attribute
}
