% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kwargs}
\alias{get_kwargs}
\title{get keyword argument values}
\usage{
get_kwargs(kwarg, kwargs, default = NULL)
}
\arguments{
\item{kwarg}{char with length of 1.}

\item{kwargs}{list object, usually passed from ...in a function.}

\item{default}{default value if no matching. default is NULL.}
}
\value{
value of either matched argument or default value if no match.
}
\description{
retreive the value from of list of key word arguments
}
\examples{
# foo <- function(...){
  params <- list(...)
  c <- get_kwargs("c", params, NULL)
  d <- get_kwargs("d", params, NULL)
  print(c)
  print(d)
}
foo(a = 1,c=4 ,b=2)
}
