% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mtx.R
\name{mtx_list}
\alias{mtx_list}
\title{Generate a Matrix List}
\usage{
mtx_list(list, group)
}
\arguments{
\item{list}{A list containing the data to be processed.}

\item{group}{A grouping variable used to organize the data in the list.}
}
\value{
A list of matrices organized by the specified grouping.
}
\description{
This function creates a matrix list based on the provided input list and grouping variable.
}
