% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrive}
\alias{retrive}
\title{Get value returned from a R process and set the default return value if
nothing is returned}
\usage{
retrive(x, default = "")
}
\arguments{
\item{x}{any R object.}

\item{default}{object returned if x is empty.}
}
\value{
either x or default value
}
\description{
Get value returned from a R process and set the default return value if
nothing is returned
}
\examples{
# ADD_EXAMPLES_HERE
retrive(NA, 1)
}
