% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_empty}
\alias{rm_empty}
\title{Remove Empty Elements from a Vector}
\usage{
rm_empty(x)
}
\arguments{
\item{x}{A character vector from which empty elements should be removed.}
}
\value{
A character vector with empty elements removed.
}
\description{
This function removes empty elements (e.g., \code{""}) from a character vector.
}
\examples{
rm_empty(c("a", "", "b", ""))
# [1] "a" "b"
}
