% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{section_cat}
\alias{section_cat}
\title{Create a Section for Categorical Variables}
\usage{
section_cat(varname, var_levels, group)
}
\arguments{
\item{varname}{A character string representing the name of the variable.}

\item{var_levels}{A vector of levels for the categorical variable.}

\item{group}{A group object.}
}
\value{
A structured section for the categorical variable.
}
\description{
This function generates a section for categorical variables in a report or analysis.
}
\examples{
section_cat("BMI", c("Underweight", "Normal", "Overweight", "Obese"), arms)
section_cat("Sex", c("Male", "Female"), arms)
}
