% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{section_grp}
\alias{section_grp}
\title{Create a Group Section}
\usage{
section_grp(group, select = NULL)
}
\arguments{
\item{group}{A group object.}

\item{select}{An optional vector specifying a display options under group header. Options are \code{c("n_title","unit")} to disaply number of subjects in each group or the units of groups. Default is \code{NULL}.}
}
\value{
A section group object (details depend on implementation).
}
\description{
This function generates a section group based on the provided variable name,
its levels, and optional units.
}
\examples{
# Example usage:
arms <- gen_grp("Treatment", c("SAR1234", "ABC", "Placebo"))
section_grp(arms, select = "n_title")
section_grp(arms, select = "unit")
}
