% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{section_list}
\alias{section_list}
\title{Generate a listing sections.}
\usage{
section_list(varname, n, group)
}
\arguments{
\item{varname}{A character string representing the name of the variable.}

\item{n}{An integer specifying the number of items to generate.}

\item{group}{A group object.}
}
\value{
A listing section.
}
\description{
This function creates a listing section using the provided variable name,
number of subjects, and group information.
}
\examples{
# generate listing for 3 sections for trial arm groups
arms <- gen_grp("Treatment", c("SAR1234", "ABC", "Placebo"))
section_list("Patient", 3, arms)
}
