% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{section_num}
\alias{section_num}
\title{Generate Numeric Section}
\usage{
section_num(varname, group, extra_setups = list(params = NULL, fmt = NULL))
}
\arguments{
\item{varname}{A character string representing the variable name.}

\item{group}{A group object.}
}
\value{
section content for numeric variables
}
\description{
This function generates a section of numeric variables based on the provided variable name and group.
}
\examples{
arms <- gen_grp("Treatment", c("SAR1234", "ABC", "Placebo"))
section_num("fAAT", arms)
}
