% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styler.R
\name{text_styler}
\alias{text_styler}
\title{Apply Text Styling}
\usage{
text_styler(x, ...)
}
\arguments{
\item{x}{A character vector or text input to be styled.}

\item{...}{Additional arguments for styling options.
\itemize{
\item \code{font_size} (default = 10): The font size for the text.
\item \code{header_font_size} (default = 10): The font size for the header text.
\item \code{footer_font_size} (default = 9): The font size for the footer text.
\item \code{footer_font} (default = "Arial Narrow"): The font for the footer text.
\item \code{body_font} (default = "Times New Roman"): The font for the body text.
}}
}
\value{
A styled text object or character vector.
}
\description{
This function applies styling to the table text based on the provided arguments.
}
\examples{
# Example usage:
tab <- data.frame(A= rep(c("A", "B"), 5), B = rep(c("C", "D"), 5))
output <- flextable(tab)
output \%>\%
text_styler(font_size = 12, body_font = "Arial Narrow")
}
